Reduce duplicate code in pocl_<pthread/tbb>_init
- always use PoCL vendor string and ID
