+++ lib/CL/devices/tbb/tbb.cc
@@ -42,12 +42,10 @@
 #include "common.h"
 #include "common_utils.h"
 #include "config.h"
-#include "cpuinfo.h"
 #include "pocl_mem_management.h"
 #include "pocl_util.h"
 #include "tbb.h"
 #include "tbb_scheduler.h"
-#include "topology/pocl_topology.h"
 
 #ifndef HAVE_LIBDL
 #error tbb driver requires DL library
@@ -57,16 +55,10 @@
 #include "pocl_llvm.h"
 #endif
 
-struct data {
-  /* Currently loaded kernel. */
-  cl_kernel current_kernel;
-  volatile uint64_t total_cmd_exec_time;
-};
-
 void
 pocl_tbb_init_device_ops(struct pocl_device_ops *ops)
 {
-  pocl_pthread_init_device_ops(ops);
+  pocl_pthread_init_device_ops (ops);
 
   ops->device_name = "tbb";
 
@@ -83,73 +75,21 @@ char scheduler_initialized = 0;
 cl_int
 pocl_tbb_init (unsigned j, cl_device_id device, const char* parameters)
 {
-  struct data *d;
-  cl_int ret = CL_SUCCESS;
-  int err;
-
-  d = (struct data *) calloc (1, sizeof (struct data));
-  if (d == NULL)
-    return CL_OUT_OF_HOST_MEMORY;
-
-  d->current_kernel = NULL;
-  device->data = d;
-
-  pocl_init_default_device_infos (device);
-  device->extensions = HOST_DEVICE_EXTENSIONS;
-
-  device->on_host_queue_props
-      = CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE | CL_QUEUE_PROFILING_ENABLE;
-
-  /* full memory consistency model for atomic memory and fence operations
-  except CL_DEVICE_ATOMIC_SCOPE_ALL_DEVICES. see
-  https://www.khronos.org/registry/OpenCL/specs/3.0-unified/html/OpenCL_API.html#opencl-3.0-backwards-compatibility*/
-  device->atomic_memory_capabilities = CL_DEVICE_ATOMIC_ORDER_RELAXED
-                                       | CL_DEVICE_ATOMIC_ORDER_ACQ_REL
-                                       | CL_DEVICE_ATOMIC_ORDER_SEQ_CST
-                                       | CL_DEVICE_ATOMIC_SCOPE_WORK_GROUP
-                                       | CL_DEVICE_ATOMIC_SCOPE_DEVICE;
-  device->atomic_fence_capabilities = CL_DEVICE_ATOMIC_ORDER_RELAXED
-                                       | CL_DEVICE_ATOMIC_ORDER_ACQ_REL
-                                       | CL_DEVICE_ATOMIC_ORDER_SEQ_CST
-                                       | CL_DEVICE_ATOMIC_SCOPE_WORK_ITEM
-                                       | CL_DEVICE_ATOMIC_SCOPE_WORK_GROUP
-                                       | CL_DEVICE_ATOMIC_SCOPE_DEVICE;
-
-  /* hwloc probes OpenCL device info at its initialization in case
-     the OpenCL extension is enabled. This causes to printout
-     an unimplemented property error because hwloc is used to
-     initialize global_mem_size which it is not yet. Just put
-     a nonzero there for now. */
-  device->global_mem_size = 1;
-  err = pocl_topology_detect_device_info (device);
-  if (err)
-    ret = CL_INVALID_DEVICE;
+  cl_int ret = pocl_device_init_common (device);
 
   /* device->max_compute_units was set up by topology_detect,
      but we use the TBB library (result should be the same).
      task_area initialization is optional and max_concurrency
      can be retrieved without prior initialization. */
   tbb::task_arena ta;
-  //ta.initialize();
-  device->max_compute_units = ta.max_concurrency();
-
-  pocl_cpuinfo_detect_device_info(device);
-  pocl_set_buffer_image_limits(device);
-
-  /* in case hwloc doesn't provide a PCI ID, let's generate
-     a vendor id that hopefully is unique across vendors. */
-  const char *magic = "pocl";
-  if (device->vendor_id == 0)
-    device->vendor_id =
-      magic[0] | magic[1] << 8 | magic[2] << 16 | magic[3] << 24;
-
-  device->vendor_id += j;
+  //ta.initialize ();
+  device->max_compute_units = ta.max_concurrency ();
 
   if (!scheduler_initialized)
     {
       scheduler_initialized = 1;
-      pocl_init_dlhandle_cache();
-      pocl_init_kernel_run_command_manager();
+      pocl_init_dlhandle_cache ();
+      pocl_init_kernel_run_command_manager ();
       tbb_scheduler_init (device);
     }
   /* system mem as global memory */
